Added a hello world source file to get things started.
Poked around the internet for some C++ unit testing frameworks and settled on Google Test.
Added what I believe to be the necessary files to the lib folder and made a crude makefile for running the tests.
Of course, all of this is moot until I get SOMETHING to test in the first place.