Multiple changes...
Adding more const where I can.

Changed any plain void methods to return the ArrayList being acted on.
This allows for a little bit of method chaining if desired.
Added a test that demonstrates this.

Added a consumeIterator function that will fully read and destroy the passed in iterator.
I'm not sure how I feel about a method freeing a passed in pointer, but this will allow for easy one line conversions between collections.