Many changes
Migrating to a newer version of GTest.
I'm now downloading GTest as a part of the build instead of committing it to the repo directly.

Creating a new version of ArrayList that only contains `int`s rather than `void*`.
Future containers will be `int` typed as well.
Also moving away from the Iterator implementation.
This is to help me focus on the datastructures and algorithms rather than worrying about keeping things data agnostic.