Misc cleanup
Explicitly setting the ArrayList default constructor to be a no-arg
method.

Cleaning up the ArrayList merge sort implementation a little bit.

Cleaning up the makefile.
The makefile was mistakenly compiling every source file for every
individual output file.
This caused `make` to think compilation was successful as long as one
file in the project compiled successfully.